refactor(foundation-demo): move header and sidebar out of misc directory
move header and sidebar out of misc directory
