+++ apps/foundation-demo/src/app/header/header.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
-import { DaffHeaderModule } from '../../design/molecules/daff-header/daff-header.module';
+import { DaffHeaderModule } from '../design/molecules/daff-header/daff-header.module';
 import { HeaderComponent } from './components/header/header.component';
 import { SidebarModule } from '../sidebar/sidebar.module';
 import { RouterModule } from '@angular/router';
