+++ apps/foundation-demo/src/app/sidebar/components/sidebar-view/sidebar-view.component.ts
