+++ apps/foundation-demo/src/app/sidebar/components/sidebar/sidebar.component.scss
