+++ apps/foundation-demo/src/app/sidebar/containers/sidebar.component.ts
