+++ apps/foundation-demo/src/app/sidebar/reducers/index.spec.ts
