+++ apps/foundation-demo/src/app/sidebar/reducers/sidebar.reducer.spec.ts
